+++ ProjetoRPG.Domain/Game/Inventory.cs
@@ -6,7 +6,7 @@ namespace ProjetoRPG.Domain.Game;
 public class Inventory : BaseEntity
 {
     public int Capacity { get; set; }
-    public double Gold { get; set; }
+    public float Gold { get; set; }
     
     //Equiped Items
     public int? IdEquippedArmor { get; set; }
